backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4
gunicorn
psycopg2-binary
PyJWT #used to create and verify tokens
bcrypt # used for password hashing , Hash is a oneway function 
# so password(register)-> hashed , password(login )->hashed and than match hashed passwords , never plain
